requirements.txt
dbt-core==1.8.7
dbt-snowflake==1.8.3
dagster-cloud==1.8.10
dagster==1.8.10
dagster-cloud-cli==1.8.10
dagster-airbyte==0.24.10
dagster-dbt==0.24.10
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
matplotlib==3.9.2
sqlalchemy==2.0.35
imbalanced-learn==0.12.4
snowflake-sqlalchemy==1.6.1
dagster-snowflake==0.24.10
snowflake-connector-python==3.12.2